complete lab2

+++ lab2/DebugPractice/DebugExercise1.java
@@ -4,7 +4,7 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        float quotient = Float.valueOf(top) / bottom;
         int result = Math.round(quotient);
         return result;
     }
